Implemented Confirmation Pop up on Delete Record.

+++ src/app/theme/pages/default/features/master_management/academic_year/academic-year-list/academic-year-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs/Rx';
 
+import { ConfirmationService } from 'primeng/primeng';
 import { GlobalErrorHandler } from '../../../../../../../_services/error-handler.service';
 import { MessageService } from '../../../../../../../_services/message.service';
 import { AcademicYear } from "../../../../_models/index";
@@ -49,6 +50,7 @@ export class AcademicYearListComponent implements OnInit {
         private messageService: MessageService,
         private academicYearService: AcademicYearService,
         private globalErrorHandler: GlobalErrorHandler,
+        private confirmationService: ConfirmationService,
         private _script: ScriptLoaderService) {
     }
 
@@ -107,14 +109,26 @@ export class AcademicYearListComponent implements OnInit {
         this.router.navigate(['/features/academicYear/edit', academicYear.id]);
     }
     onAcademicYearDeleteClick(academicYear: AcademicYear) {
-        this.academicYearService.deleteAcademicYear(academicYear.id).subscribe(
-            results => {
-                this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Deleted Successfully' });
-                this.getAllAcademicYears();
+        this.confirmationService.confirm({
+            message: 'Do you want to delete this record?',
+            header: 'Delete Confirmation',
+            icon: 'fa fa-trash',
+            accept: () => {
+                this.academicYearService.deleteAcademicYear(academicYear.id).subscribe(
+                    results => {
+                        this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Deleted Successfully' });
+                        if ((this.currentPageNumber - 1) * this.perPage == (this.total - 1)) {
+                            this.currentPageNumber--;
+                        }
+                        this.getQueryDataCount();
+                    },
+                    error => {
+                        this.globalErrorHandler.handleError(error);
+                    })
             },
-            error => {
-                this.globalErrorHandler.handleError(error);
-            })
+            reject: () => {
+            }
+        });
     }
     onAddAcademicYear() {
         this.router.navigate(['/features/academicYear/add']);
